requirements.txt
aws-cdk-lib==2.81.0
constructs>=10.0.0,<11.0.0
attrs==23.1.0
aws-cdk-lib==2.93.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.166
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.aws-kinesisfirehose-alpha==2.93.0a0
aws-cdk.aws-kinesisfirehose-destinations-alpha==2.93.0a0
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
constructs==10.2.69
exceptiongroup==1.1.3
idna==3.4
importlib-resources==6.0.1
jsii==1.87.0
oauthlib==3.2.2
ondilo==0.3.0
publication==0.0.3
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typeguard==2.13.3
typing_extensions==4.7.1
urllib3==2.0.4
